requirements.txt
aiofiles==24.1.0
aiohttp==3.10.5
ffmpeg-python==0.2.0
g4f[all]==0.3.2.4
jaconv==0.4.0
mutagen==1.47.0
openai==1.41.0
pillow==10.4.0
py-cord[voice]==2.6.0
pyshorteners==1.0.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
youtube-search-python==1.6.6
yt-dlp==2024.8.6